PortOps AI/PortOps AI/requirements.txt
# Core web framework
flask>=2.3.0

# Environment management
python-dotenv>=1.0.0

# Google AI/Gemini
google-generativeai>=0.3.0

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.0.1

# PDF processing
PyPDF2>=3.0.0

# Vector store and ML
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Web utilities (comes with Flask but good to be explicit)
werkzeug>=2.3.0
gunicorn
python-docx
